+++ packages/cli/src/utils/files.js
@@ -207,9 +207,14 @@ function* walkDir(dir) {
   const entries = fs.readdirSync(dir, { withFileTypes: true });
   for (const entry of entries) {
     if (entry.isDirectory()) {
-      const subDir = path.join(entry.parentPath, entry.name);
+      const subDir = path.join(dir, entry.name);
       yield* walkDir(subDir);
     } else if (entry.isFile() || entry.isSymbolicLink()) {
+      if (!entry.parentPath) {
+        // dirent.parentPath is only available since Node.js 18.20, 20.12, and
+        // 21.4. Re-assigning it so the caller can use dirent.parentPath.
+        entry.parentPath = dir;
+      }
       yield entry;
     }
   }
@@ -223,10 +228,15 @@ function* walkDirLimitedLevels(dir, maxLevels, currentLevel = 1) {
   for (const entry of entries) {
     if (entry.isDirectory()) {
       if (currentLevel < maxLevels) {
-        const subDir = path.join(entry.parentPath, entry.name);
+        const subDir = path.join(dir, entry.name);
         yield* walkDirLimitedLevels(subDir, maxLevels, currentLevel + 1);
       }
     } else if (entry.isFile() || entry.isSymbolicLink()) {
+      if (!entry.parentPath) {
+        // dirent.parentPath is only available since Node.js 18.20, 20.12, and
+        // 21.4. Re-assigning it so the caller can use dirent.parentPath.
+        entry.parentPath = dir;
+      }
       yield entry;
     }
   }